Move validate_api to util and review fixes
